requirements.txt
python-telegram-bot==21.9
httpx==0.27.2
pydantic>=2.12.5
python-dotenv==1.0.1
json-repair==0.27.0
loguru==0.7.2
beautifulsoup4==4.12.3
